MqttHandler/app/requirementsMqtt.txt
redis == 4.6.0
simplejson == 3.17.6
paho-mqtt == 1.6.1